app/requirements.txt
langchain==0.0.347
langchain-experimental==0.0.44
openai==1.3.7
botbuilder-integration-aiohttp>=4.14.4
faiss-cpu
tiktoken
docx2txt
pillow
pypdf
tenacity
tabulate
azure-cosmos
streamlit
python-dotenv
azure-ai-formrecognizer
azure-storage-blob
rarfile
setuptools